+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Head
+# 3.1.3
 
 * Fixed: Bug causing preventing the disabling of rules analyzing the `root` node, including: `max-line-length`, `max-empty-lines`, `no-eol-whitespace`, `no-missing-eof-newline`, and `string-quotes`.
 * Fixed: Bug causing `rule-properties-order` to get confused by properties with an unspecified order.